Refactor BdGame class to encapsulate XLSM parsing logic and move parsing functionality from parse_data.py
